Chore: Reconstruct backend&model dir structure
