+++ backend/README.md
@@ -1,3 +1,9 @@
+1.
+```
+bash backend_env_setup.sh
+```
+
+2.
 ```
 python main.py
 ```
