--- backend/backend_cpu_env_setup.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-pip install -r requirements.txt
-pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
