+++ backend/langgraph_/utils.py
@@ -1,7 +1,4 @@
 from langchain_core.runnables import RunnableConfig
-from transformers import AutoModelForCausalLM, AutoTokenizer
-from transformers import BitsAndBytesConfig
-from huggingface_hub import login
 import argparse, os, re, csv
 from datetime import datetime
 
@@ -64,40 +61,6 @@ def str2bool(v):
         return False
     else:
         raise argparse.ArgumentTypeError("Boolean value expected.")
-
-
-def load_qwen_model():
-    from unsloth import FastLanguageModel
-
-    global model, tokenizer
-    # Hugging Face 토큰 설정
-
-    # 환경 변수에서 Hugging Face 토큰 가져오기
-    huggingface_token = os.getenv("HUGGINGFACE_TOKEN")
-    login(token=huggingface_token)
-
-    # 모델 이름 설정
-    model_name = "unsloth/Qwen2.5-Coder-32B-Instruct"
-
-    # 8비트 양자화 설정
-    bnb_config = BitsAndBytesConfig(load_in_8bit=True)
-
-    home_path = os.path.expanduser("~")
-    cache_dir = os.path.join(home_path, "sql-helper/.cache/unsloth")
-
-    # FastLanguageModel을 사용하여 모델과 토크나이저 로드
-    model, tokenizer = FastLanguageModel.from_pretrained(
-        model_name=model_name,
-        token=huggingface_token,
-        quantization_config=bnb_config,
-        cache_dir=cache_dir,
-        device_map="auto",
-    )
-
-    # Unsloth 사용 시 inference 모드 전환
-    model = FastLanguageModel.for_inference(model)
-
-    return model, tokenizer
 
 
 def extract_context_tables(table_contexts, table_contexts_ids):
