+++ frontend/README.md
@@ -1,3 +1,9 @@
+1.
+```
+bash frontend_env_setup.sh
+```
+
+2.
 ```
 streamlit run app.py
 ```
